migrate text jest tests to vitest
